Add support for removing keys at the end of a session.
These changes track the number of concurrent sessions and at the end of the
last session removes all active keys.  The code currently assumes that only one
key is used for login but supports additional keys that are added after login.
It does not support the user directory being remounted after login.