Determine stale session counter files based on system boot time.
Previously, a stale session counter file was based on whether the passphrase
had been used previously.  This, however, assumes that only one passphrase is
used, which in the case of pefs, is incorrect.

The most common case causing a corrupt session file is for a reboot that leaves
the session counter file with a value above 0.  To handle this case the
modification time of the session counter file is compared against the boot time.
If it was modified before the last boot, it is forced to have a value of 0.