If user's home directory is a valid pefs mount then always count the session.
This change allows the pam_pefs session functionality to be used independently
of the pam_pefs authentication functionality.  One particular use case is with
an ssh login, where the user adds keys.  In this case the keys will be cleared
as expected if pam_pefs is included in the ssh session.